+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
- - feat(nextjs,remix,sveltekit): add support to add MCP server during configure ([#1063](https://github.com/getsentry/sentry-wizard/pull/1063))
+- feat(nextjs,remix,sveltekit): add support to add MCP server during configure ([#1063](https://github.com/getsentry/sentry-wizard/pull/1063))
+- fix(nextjs): Prevent double-wrapping of Next.js config ([#1058](https://github.com/getsentry/sentry-wizard/pull/1058))
+- feat(flutter): Add structured logs opt-in #1050 ([#1050](https://github.com/getsentry/sentry-wizard/pull/1050))
 
 ## 6.1.2